+++ etna/metrics/metrics.py
@@ -1,7 +1,7 @@
 from functools import partial
 
 from etna.metrics.base import Metric
-from etna.metrics.base import MetricAggregationMode
+from etna.metrics.base import MetricWithMissingHandling
 from etna.metrics.functional_metrics import mae
 from etna.metrics.functional_metrics import mape
 from etna.metrics.functional_metrics import max_deviation
@@ -26,7 +26,7 @@ class MAE(Metric):
     You can read more about logic of multi-segment metrics in Metric docs.
     """
 
-    def __init__(self, mode: str = MetricAggregationMode.per_segment, **kwargs):
+    def __init__(self, mode: str = "per-segment", **kwargs):
         """Init metric.
 
         Parameters
@@ -45,29 +45,47 @@ class MAE(Metric):
         return False
 
 
-class MSE(Metric):
+class MSE(MetricWithMissingHandling):
     """Mean squared error metric with multi-segment computation support.
 
     .. math::
         MSE(y\_true, y\_pred) = \\frac{\\sum_{i=1}^{n}{(y\_true_i - y\_pred_i)^2}}{n}
+
+    This metric can handle missing values with parameter ``missing_mode``.
+    If there are too many of them in ``ignore`` mode, the result will be ``None``.
 
     Notes
     -----
     You can read more about logic of multi-segment metrics in Metric docs.
     """
 
-    def __init__(self, mode: str = MetricAggregationMode.per_segment, **kwargs):
+    def __init__(self, mode: str = "per-segment", missing_mode: str = "error", **kwargs):
         """Init metric.
 
         Parameters
         ----------
-        mode: 'macro' or 'per-segment'
-            metrics aggregation mode
+        mode:
+            "macro" or "per-segment", way to aggregate metric values over segments:
+
+            * if "macro" computes average value
+
+            * if "per-segment" -- does not aggregate metrics
+
+            See :py:class:`~etna.metrics.base.MetricAggregationMode`.
+
+        missing_mode:
+            mode of handling missing values (see :py:class:`~etna.metrics.base.MetricMissingMode`)
         kwargs:
             metric's computation arguments
         """
         mse_per_output = partial(mse, multioutput="raw_values")
-        super().__init__(mode=mode, metric_fn=mse_per_output, metric_fn_signature="matrix_to_array", **kwargs)
+        super().__init__(
+            mode=mode,
+            metric_fn=mse_per_output,
+            missing_mode=missing_mode,
+            metric_fn_signature="matrix_to_array",
+            **kwargs,
+        )
 
     @property
     def greater_is_better(self) -> bool:
@@ -86,13 +104,19 @@ class RMSE(Metric):
     You can read more about logic of multi-segment metrics in Metric docs.
     """
 
-    def __init__(self, mode: str = MetricAggregationMode.per_segment, **kwargs):
+    def __init__(self, mode: str = "per-segment", **kwargs):
         """Init metric.
 
         Parameters
         ----------
-        mode: 'macro' or 'per-segment'
-            metrics aggregation mode
+        mode:
+            "macro" or "per-segment", way to aggregate metric values over segments:
+
+            * if "macro" computes average value
+
+            * if "per-segment" -- does not aggregate metrics
+
+            See :py:class:`~etna.metrics.base.MetricAggregationMode`.
         kwargs:
             metric's computation arguments
         """
@@ -115,13 +139,19 @@ class R2(Metric):
     You can read more about logic of multi-segment metrics in Metric docs.
     """
 
-    def __init__(self, mode: str = MetricAggregationMode.per_segment, **kwargs):
+    def __init__(self, mode: str = "per-segment", **kwargs):
         """Init metric.
 
         Parameters
         ----------
-        mode: 'macro' or 'per-segment'
-            metrics aggregation mode
+        mode:
+            "macro" or "per-segment", way to aggregate metric values over segments:
+
+            * if "macro" computes average value
+
+            * if "per-segment" -- does not aggregate metrics
+
+            See :py:class:`~etna.metrics.base.MetricAggregationMode`.
         kwargs:
             metric's computation arguments
         """
@@ -145,13 +175,19 @@ class MAPE(Metric):
     You can read more about logic of multi-segment metrics in Metric docs.
     """
 
-    def __init__(self, mode: str = MetricAggregationMode.per_segment, **kwargs):
+    def __init__(self, mode: str = "per-segment", **kwargs):
         """Init metric.
 
         Parameters
         ----------
-        mode: 'macro' or 'per-segment'
-            metrics aggregation mode
+        mode:
+            "macro" or "per-segment", way to aggregate metric values over segments:
+
+            * if "macro" computes average value
+
+            * if "per-segment" -- does not aggregate metrics
+
+            See :py:class:`~etna.metrics.base.MetricAggregationMode`.
         kwargs:
             metric's computation arguments
         """
@@ -175,13 +211,19 @@ class SMAPE(Metric):
     You can read more about logic of multi-segment metrics in Metric docs.
     """
 
-    def __init__(self, mode: str = MetricAggregationMode.per_segment, **kwargs):
+    def __init__(self, mode: str = "per-segment", **kwargs):
         """Init metric.
 
         Parameters
         ----------
-        mode: 'macro' or 'per-segment'
-            metrics aggregation mode
+        mode:
+            "macro" or "per-segment", way to aggregate metric values over segments:
+
+            * if "macro" computes average value
+
+            * if "per-segment" -- does not aggregate metrics
+
+            See :py:class:`~etna.metrics.base.MetricAggregationMode`.
         kwargs:
             metric's computation arguments
         """
@@ -205,13 +247,19 @@ class MedAE(Metric):
     You can read more about logic of multi-segment metrics in Metric docs.
     """
 
-    def __init__(self, mode: str = MetricAggregationMode.per_segment, **kwargs):
+    def __init__(self, mode: str = "per-segment", **kwargs):
         """Init metric.
 
         Parameters
         ----------
-        mode: 'macro' or 'per-segment'
-            metrics aggregation mode
+        mode:
+            "macro" or "per-segment", way to aggregate metric values over segments:
+
+            * if "macro" computes average value
+
+            * if "per-segment" -- does not aggregate metrics
+
+            See :py:class:`~etna.metrics.base.MetricAggregationMode`.
         kwargs:
             metric's computation arguments
         """
@@ -235,13 +283,19 @@ class MSLE(Metric):
     You can read more about logic of multi-segment metrics in Metric docs.
     """
 
-    def __init__(self, mode: str = MetricAggregationMode.per_segment, **kwargs):
+    def __init__(self, mode: str = "per-segment", **kwargs):
         """Init metric.
 
         Parameters
         ----------
-        mode: 'macro' or 'per-segment'
-            metrics aggregation mode
+        mode:
+            "macro" or "per-segment", way to aggregate metric values over segments:
+
+            * if "macro" computes average value
+
+            * if "per-segment" -- does not aggregate metrics
+
+            See :py:class:`~etna.metrics.base.MetricAggregationMode`.
         kwargs:
             metric's computation arguments
 
@@ -266,13 +320,19 @@ class Sign(Metric):
     You can read more about logic of multi-segment metrics in Metric docs.
     """
 
-    def __init__(self, mode: str = MetricAggregationMode.per_segment, **kwargs):
+    def __init__(self, mode: str = "per-segment", **kwargs):
         """Init metric.
 
         Parameters
         ----------
-        mode: 'macro' or 'per-segment'
-            metrics aggregation mode
+        mode:
+            "macro" or "per-segment", way to aggregate metric values over segments:
+
+            * if "macro" computes average value
+
+            * if "per-segment" -- does not aggregate metrics
+
+            See :py:class:`~etna.metrics.base.MetricAggregationMode`.
         kwargs:
             metric's computation arguments
         """
@@ -296,13 +356,19 @@ class MaxDeviation(Metric):
     You can read more about logic of multi-segment metrics in Metric docs.
     """
 
-    def __init__(self, mode: str = MetricAggregationMode.per_segment, **kwargs):
+    def __init__(self, mode: str = "per-segment", **kwargs):
         """Init metric.
 
         Parameters
         ----------
-        mode: 'macro' or 'per-segment'
-            metrics aggregation mode
+        mode:
+            "macro" or "per-segment", way to aggregate metric values over segments:
+
+            * if "macro" computes average value
+
+            * if "per-segment" -- does not aggregate metrics
+
+            See :py:class:`~etna.metrics.base.MetricAggregationMode`.
         kwargs:
             metric's computation arguments
         """
@@ -320,18 +386,25 @@ class WAPE(Metric):
 
     .. math::
         WAPE(y\_true, y\_pred) = \\frac{\\sum_{i=1}^{n} |y\_true_i - y\_pred_i|}{\\sum_{i=1}^{n}|y\\_true_i|}
+
     Notes
     -----
     You can read more about logic of multi-segment metrics in Metric docs.
     """
 
-    def __init__(self, mode: str = MetricAggregationMode.per_segment, **kwargs):
+    def __init__(self, mode: str = "per-segment", **kwargs):
         """Init metric.
 
         Parameters
         ----------
-        mode: 'macro' or 'per-segment'
-            metrics aggregation mode
+        mode:
+            "macro" or "per-segment", way to aggregate metric values over segments:
+
+            * if "macro" computes average value
+
+            * if "per-segment" -- does not aggregate metrics
+
+            See :py:class:`~etna.metrics.base.MetricAggregationMode`.
         kwargs:
             metric's computation arguments
         """